[ECP-8602] Define guest endpoints for giftcard implementation
